@@ -12,14 +12,15 @@
     "kabelwerk": ".."
   },
   "devDependencies": {
-    "parcel": "^2.3.1"
+    "parcel": "^2.6.2",
+    "process": "^0.11.10"
   },
   "dependencies": {
-    "evergreen-ui": "^6.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "evergreen-ui": "^6.10.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "scripts": {
-    "start": "parcel --cache-dir .parcel-cache"
+    "start": "parcel"
   }
 }
