@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { Chat } from './Chat.jsx';
 import { KabelwerkProvider } from './KabelwerkContext.jsx';
@@ -32,4 +32,4 @@ const App = () => {
     );
 };
 
-ReactDOM.render(<App />, document.getElementById('app'));
+createRoot(document.getElementById('app')).render(<App />);
